TST: Removed `comm_self` fixture
